Update iostat.md
iostat的util参数解读有误。参照于man手册
Percentage of elapsed time during which I/O requests were issued to the device (bandwidth utilization for the device). Device saturation occurs when this value is close to 100% for  devices  serving  requests  serially.But  for  devices serving requests in parallel, such as RAID arrays and modern SSDs, this number does not reflect their performance limits.